% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{AddMarkers}
\alias{AddMarkers}
\title{Add markers to celllabeler ude results and perform prediction}
\usage{
AddMarkers(res, counts, markers, cluster.id)
}
\arguments{
\item{res}{A celllabeler output list involoving at least ModelFits}

\item{counts}{A normalized gene expression matrix}

\item{markers}{Marker list with name being cell type names}

\item{cluster.id}{Character vectors of cluster information for each cell.}
}
\value{
A celllabeler output list involoving predictions
}
\description{
Add markers to celllabeler ude results and perform prediction
}
\author{
Jin ning
}
