% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celllabeler.object.R
\name{AddPredictionToMeta}
\alias{AddPredictionToMeta}
\title{Add the prediction results to meta.data in CellLabelerObject for plot use}
\usage{
AddPredictionToMeta(object, cluster.var, prediction.var = "prediction")
}
\arguments{
\item{object}{A cellLabeler object with predictions}

\item{cluster.var}{Column name in meta data of the object indicating cluster information}

\item{prediction.var}{A new defined column name in meta data of the object indicating cluster information}
}
\value{
A cellLabeler object
}
\description{
Add the prediction results to meta.data in CellLabelerObject for plot use
}
