% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{allGenes}
\alias{allGenes}
\title{Collect all genes in the analysis,}
\usage{
allGenes(de_model)
}
\arguments{
\item{de_model}{Results of FindAllUniqueMarkers().}
}
\description{
Collect all genes in the analysis,
}
