% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclustering.R
\name{AnnoySearch}
\alias{AnnoySearch}
\title{Search an Annoy approximate nearest neighbor index}
\usage{
AnnoySearch(index, query, k, search.k = -1, include.distance = TRUE)
}
\arguments{
\item{query}{A set of data to be queried against the index}

\item{k}{Number of neighbors}

\item{search.k}{During the query it will inspect up to search_k nodes which gives you a run-time tradeoff between better accuracy and speed.}

\item{include.distance}{Include the corresponding distances in the result}

\item{Annoy}{index, built with AnnoyBuildIndex}
}
\value{
A list with 'nn.idx' (for each element in 'query', the index of the nearest k elements in the index) and 'nn.dists' (the distances of the nearestk elements)
}
\description{
Search an Annoy approximate nearest neighbor index
}
