% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclustering.R
\name{as.WebGraph.Neighbor}
\alias{as.WebGraph.Neighbor}
\title{As web graph functions when given a neighbor}
\usage{
\method{as.WebGraph}{Neighbor}(x, weighted = TRUE, ...)
}
\description{
As web graph functions when given a neighbor
}
