% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celllabeler.object.R
\docType{class}
\name{CellLabeler-class}
\alias{CellLabeler-class}
\title{Each CellLabeler object has a number of slots which store information. Key slots to access are listed below.}
\description{
Each CellLabeler object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{counts}}{The raw expression count matrix}

\item{\code{data}}{Normalized gene matrix}

\item{\code{meta.data}}{Cell-level metadata}

\item{\code{ModelFits}}{CellLabeler model fitting results}

\item{\code{ude}}{CellLabeler detected uniquely differential expression genes}

\item{\code{prediction}}{CellLabeler prediction dataframe}

\item{\code{num.core}}{The number of core used in the package}
}}

