% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celllabeler.object.R
\name{show,CellLabeler-method}
\alias{show,CellLabeler-method}
\title{Overview of a Celllabeler object}
\usage{
\S4method{show}{CellLabeler}(object)
}
\arguments{
\item{object}{A celllabeler object}
}
\value{
\code{show}: Prints summary to \code{\link[base]{stdout}} and invisibly returns \code{NULL}
}
\description{
Overview of a Celllabeler object
}
\section{Functions}{
\itemize{
\item \code{show(CellLabeler)}: Overview of a \code{CellLabeler} object

}}
