% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLabeler.R
\name{celllabeler.default}
\alias{celllabeler.default}
\title{Fitting the constrained spline model to perform temporal differential expression analysis for time course scRNA-seq data}
\usage{
\method{celllabeler}{default}(
  object,
  markers = NULL,
  sample.id = NULL,
  cluster.id = NULL,
  similar.gene = 20,
  similar.pct = 0.8,
  lfc = 0.5,
  min.ccells = 10,
  max.genes = 100,
  mod = "combine",
  up.thr = 0.9,
  num.core = 1,
  verbose = FALSE
)
}
\arguments{
\item{object}{Gene expression raw count matrix}

\item{sample.id}{Column names of sample information in metadata.}

\item{cluster.id}{Column names of cluster information in metadata.}

\item{similar.gene}{Integer value as the number of top expressed genes in considering combining clusters before DE test}

\item{similar.pct}{Numeric value in \link{0,1} as the threshold to combine clusters before DE test when either two cluster have enough overlapped highly expressed genes. Default is 0.8.}

\item{lfc}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between the interested clusters and the rest others. Default is 0.5.}

\item{min.ccells}{Minimum number of cells in each cluster required. Default is 10.}

\item{max.genes}{Maximum number of highly expressed genes in each cluster required. Default is 100.}

\item{mod}{Character as "combine" or "all", specifying if combining similar clusters or not before DE test. Default is "combine"}

\item{up.thr}{Numberic value in \link{0,1} as the proportion of the remaining clusters in which a gene is upregulated in the interested group compared with it. Highter, the more strict. Default 0.9.}

\item{num.core}{Number of cores in multi-threaded running. Default is 1.}
}
\description{
Fitting the constrained spline model to perform temporal differential expression analysis for time course scRNA-seq data
}
\examples{

data(exampledata)
res <- celllabeler(object=counts, sample.id = sample.id, cluster.id = cluster.id, markers = markers, num.core = 10)

}
