% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{CellLabeler.plot}
\alias{CellLabeler.plot}
\title{Plot gene score for clusters regarding each possible cell types.}
\usage{
CellLabeler.plot(model_predict)
}
\arguments{
\item{model_predict}{Results of ComputePrediction(). A list contating scores.}

\item{marker_list}{List of makers, with list names being cell type names.}

\item{PCT_mat}{Gene expression percent matrix. Rows are genes and columns are clusters.}
}
\description{
Plot gene score for clusters regarding each possible cell types.
}
