% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclustering.R
\name{ClustLouvain}
\alias{ClustLouvain}
\title{we use Louvain method}
\usage{
ClustLouvain(
  object,
  modularity.fxn = 1,
  resolution = 0.8,
  algorithm = 1,
  n.start = 10,
  n.iter = 10,
  random.seed = 0,
  group.singletons = TRUE,
  temp.file.location = NULL,
  edge.file.name = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Clustering object}

\item{resolution}{Resolution parameter}

\item{verbose}{Print output}
}
\description{
we use Louvain method
}
\concept{clustering}
