% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{ComputeGeneScore}
\alias{ComputeGeneScore}
\title{Compute gene score based on combined pvalue and gene expression percents.}
\usage{
ComputeGeneScore(de_model, i.celltype, PCT_mat)
}
\arguments{
\item{de_model}{Results of FindAllUniqueMarkers().}

\item{i.celltype}{Column names specifying the interested cluster.}

\item{PCT_mat}{Gene expression percent matrix. Rows are genes and columns are clusters.}
}
\description{
Compute gene score based on combined pvalue and gene expression percents.
}
