% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{ComputePrediction}
\alias{ComputePrediction}
\title{Compute gene score and perform predictions.}
\usage{
ComputePrediction(de_model, marker_list, PCT_mat)
}
\arguments{
\item{de_model}{Results of FindAllUniqueMarkers().}

\item{marker_list}{List of makers, with list names being cell type names.}

\item{PCT_mat}{Gene expression percent matrix. Rows are genes and columns are clusters.}
}
\description{
Compute gene score and perform predictions.
}
