% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeSNN}
\alias{ComputeSNN}
\title{ComputeSNN of a given matrix}
\usage{
ComputeSNN(nn_ranked, prune)
}
\arguments{
\item{nn_ranked}{A symmetric matrix}
}
\description{
ComputeSNN of a given matrix
}
