% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celllabeler.object.R
\name{CreateCellLabelerObject}
\alias{CreateCellLabelerObject}
\title{Create the CellLabeler object with filtering step}
\usage{
CreateCellLabelerObject(
  counts,
  meta.data,
  project = "CellLabeler",
  pct.cells = 0.005,
  min.cells = 0,
  min.features = 0,
  min.umi = 10,
  num.core = 1,
  ude.result = NULL,
  res.path = NULL
)
}
\arguments{
\item{counts}{Gene expression count matrix (data.frame), p x n -- p is the number of genes and n is the number of cells}

\item{meta.data}{Meta data dataframe to store batch and clustering information}

\item{min_total_counts}{The minimum counts for each cell for filtering}

\item{percentage}{The percentage of cells that are expressed for analysis}
}
\value{
Returns CellLabeler object with filtered gene expression matrix
}
\description{
Create the CellLabeler object with filtering step
}
