% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{ExpandPrediction}
\alias{ExpandPrediction}
\title{Split the prediction dataframe when there is combination before annotation}
\usage{
ExpandPrediction(prediction)
}
\arguments{
\item{prediction}{The dataframe of cell type annotation}
}
\value{
The dataframe of cell type annotation
}
\description{
Split the prediction dataframe when there is combination before annotation
}
