% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastMapEculidean}
\alias{FastMapEculidean}
\title{FastMap for dimension reduction, SIGMOD '95: Proceedings of the 1995 ACM SIGMOD international conference on Management of data
June 1995 Pages 163–174; DOI: https://doi.org/10.1145/223784.223812}
\usage{
FastMapEculidean(D, num_dim)
}
\arguments{
\item{D}{Distance matrix}

\item{dat}{a matrix containing clustering solutions in columns}
}
\description{
Computes reduced dimension with linear computation time
}
