% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LaplacianMat}
\alias{LaplacianMat}
\title{Graph Laplacian calculation}
\usage{
LaplacianMat(A, sigma, alpha)
}
\arguments{
\item{A}{symmetric matrix}
}
\description{
Calculate graph Laplacian of a symmetrix matrix
}
