% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustering.R
\name{multiple_clustering}
\alias{multiple_clustering}
\title{Perform multiple rounds of clustering by combining HVGs selected, graph construction and Louvain clustering}
\usage{
multiple_clustering(
  counts,
  loop.max = 10,
  k = 50,
  min.cells = 100,
  num.hvg = 10,
  num.de = 10,
  logfc = 0.1,
  min.umi = 100,
  verbose = T
)
}
\arguments{
\item{counts}{A gene by cell raw count gene expression matrix}

\item{loop.max}{Integer for maximum iteration rounds; default is 10}

\item{k}{Number of the nearest neighbors in knn; default is 50}

\item{min.cells}{Minimum size of the cluster to be further spliting; default is 100}

\item{num.hvg}{Minimum number of HVGs detected for the cluster to be further spliting; default is 10}

\item{num.de}{Minimum number of DEGs detected for each update cluster; default is 10}

\item{logfc}{LogFC threshold for detecting DEGs}

\item{min.umi}{Minimum umi of genes in the filtration step}

\item{verbose}{Bool indicator for printing messages}
}
\value{
A dataframe with rownames being cell names and each column defining clustering for each round
}
\description{
Perform multiple rounds of clustering by combining HVGs selected, graph construction and Louvain clustering
}
