% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_func.R
\name{MultipleClustering}
\alias{MultipleClustering}
\title{Iterative clustering function}
\usage{
MultipleClustering(
  counts,
  hvg.method = c("poisson", "vst"),
  pca = T,
  npc = 20,
  loop.max = 10,
  k = 50,
  min.cells = 100,
  min.umi = 100,
  num.hvg = 100,
  num.vst = 3000,
  num.de = 20,
  logfc = 0.1,
  verbose = T
)
}
\arguments{
\item{counts}{Raw count gene expression matrix}

\item{hvg.method}{A charater displaying the method for HVG identification;}

\item{pca}{Bool indicator, if performing PCA before constructed SNN}

\item{npc}{An integer defining the number of top PCs for downstream analysis}

\item{loop.max}{An integer defining the maximum of rounds in iterative cluster}

\item{k}{Number of neighbors}

\item{min.cells}{The minimum threshold for clustering size}

\item{min.umi}{For filtering genes}

\item{num.vst}{Number of features in HighlyVariableGenes}

\item{num.de}{An integer threshold; cluster with DE less than this will not be subclustered}

\item{logfc}{A numeric threshold for filtering DE}

\item{verbose}{Bool indicator to print the progress}
}
\value{
A dataframe of clustering results; rows are cells and columns are rounds of clustering
}
\description{
Iterative clustering function
}
