% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parsesED2}
\alias{parsesED2}
\title{Euclidean distance calculation}
\usage{
parsesED2(X, num_core)
}
\arguments{
\item{A}{symmetric matrix}
}
\description{
Calculate graph Laplacian of a symmetrix matrix
}
