% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{perm2}
\alias{perm2}
\title{Permutation of two of all categories}
\usage{
perm2(vec)
}
\arguments{
\item{vec}{A vector of elements where my permutation performs}
}
\value{
Returns a list of permutations of two elements
}
\description{
Permutation of two of all categories
}
