% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{PlotModelScore}
\alias{PlotModelScore}
\title{Plot gene score for clusters regarding each possible cell types.}
\usage{
PlotModelScore(res)
}
\arguments{
\item{res}{A cellLabeler results with prediction}
}
\value{
A ggplot displaying prediction scores
}
\description{
Plot gene score for clusters regarding each possible cell types.
}
