% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{prediction_plot}
\alias{prediction_plot}
\title{Plot gene score for clusters regarding each possible cell types.}
\usage{
prediction_plot(object)
}
\arguments{
\item{object}{A cellLabeler object with already run celllabeler()}
}
\value{
A ggplot displaying prediction scores
}
\description{
Plot gene score for clusters regarding each possible cell types.
}
