% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{RankAllUniqueMarkers}
\alias{RankAllUniqueMarkers}
\title{Select and rank the ude to build up a list of ude for each clusters
mainly used in ude benchmark}
\usage{
RankAllUniqueMarkers(de_model, counts, cluster.id)
}
\arguments{
\item{de_model}{A list of output from FindAllUniqueMarkers()}

\item{counts}{Raw count gene expression matrix; rows are genes and columns as cells}

\item{cluster.id}{A vector of cluster labels for cells; same as input in FindAllUniqueMarkers(); if combined, then this cluster.id is the combined cluster label}
}
\value{
A list of combined statistic for each cluster
}
\description{
Select and rank the ude to build up a list of ude for each clusters
mainly used in ude benchmark
}
