% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{RegressOutMatrix}
\alias{RegressOutMatrix}
\title{Regress out the sample batch effects}
\usage{
RegressOutMatrix(
  data.expr,
  latent.data = NULL,
  features.regress = NULL,
  model.use = NULL,
  use.umi = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data.expr}{An expression matrix to regress the effects of latent.data out
of should be the complete expression matrix in genes x cells}

\item{latent.data}{A matrix or data.frame of latent variables, should be cells
x latent variables, the colnames should be the variables to regress}

\item{features.regress}{An integer vector representing the indices of the
genes to run regression on}

\item{model.use}{Model to use, one of 'linear', 'poisson', or 'negbinom'; pass
NULL to simply return data.expr}

\item{use.umi}{Regress on UMI count data}

\item{verbose}{Display a progress bar}
}
\description{
Regress out the sample batch effects
}
