% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustering.R
\name{select_hvg}
\alias{select_hvg}
\title{Detect variable genes}
\usage{
select_hvg(counts, min.pct = 0.1)
}
\arguments{
\item{counts}{A raw count gene expression matrix; rows are genes and coumns are cells}

\item{min.pct}{Numeric value threshold, genes with observed expression percent lower than it would be filtered out}
}
\value{
A character vectors of variable gene names
}
\description{
Detect variable genes
}
