% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_func.R
\name{SelectHVG}
\alias{SelectHVG}
\title{Detect variable genes with observed expression percent lower than a threshold}
\usage{
SelectHVG(counts, min.pct = 0.1)
}
\arguments{
\item{counts}{A raw count gene expression matrix; rows are genes and coumns are cells}

\item{min.pct}{Numeric value threshold, genes with observed expression percent lower than it would be filtered out}
}
\value{
A character vectors of variable gene names
}
\description{
Detect variable genes with observed expression percent lower than a threshold
}
