% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustering.R
\name{update_clustering}
\alias{update_clustering}
\title{Detect a more explicit clustering structure with a given clustering labels}
\usage{
update_clustering(
  counts,
  cluster.id,
  k = 50,
  num.hvg = 10,
  num.de = 10,
  logfc = 0.1,
  min.umi = 100,
  verbose = T
)
}
\arguments{
\item{counts}{A raw count gene expression matrix; rows are genes and columns are cells}

\item{cluster.id}{A character vectors defining the original clustering labels}

\item{k}{Number of the nearest neighbors in knn; default is 50}

\item{num.hvg}{Minimum number of HVGs detected for the cluster to be further spliting; default is 10}

\item{num.de}{Minimum number of DEGs detected for each update cluster; default is 10}

\item{logfc}{LogFC threshold for detecting DEGs}

\item{min.umi}{Minimum umi of genes in the filtration step}

\item{verbose}{Bool indicator for printing messages}
}
\value{
A list of character vectors defining the updated clustering labels and hvgs
}
\description{
Detect a more explicit clustering structure with a given clustering labels
}
