% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_func.R
\name{UpdateClustering}
\alias{UpdateClustering}
\title{Update clustering results}
\usage{
UpdateClustering(
  full.snn,
  normdata,
  cluster.id,
  num.de = 20,
  logfc = 0.1,
  verbose = T
)
}
\arguments{
\item{full.snn}{A snn network constructed from the begining}

\item{normdata}{A normalized gene expression matrix with rows being genes and columns being samples}

\item{cluster.id}{A vector of charactors displaying the different clustering labels; names of it should be cell names}

\item{num.de}{An integer threshold; cluster with DE less than this will not be subclustered}

\item{logfc}{A numeric threshold for filtering DE}

\item{verbose}{Bool indicator to print the progress}
}
\value{
A vector of charactors displaying the updated clustering labels
}
\description{
Update clustering results
}
