% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclustering.R
\docType{class}
\name{WebGraph-class}
\alias{WebGraph-class}
\alias{WebGraph}
\title{The WebGraph Class}
\description{
The WebGraph class inherits from \code{\link[Matrix:sparseMatrix]{dgCMatrix}}.
We do this to enable future expandability of graphs.
}
\section{Slots}{

\describe{
\item{\code{assay.used}}{Optional name of assay used to generate \code{WebGraph} object}
}}

\seealso{
\code{\link[Matrix]{dgCMatrix-class}}
}
