% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celllabeler.object.R
\docType{class}
\name{CellLabeler-class}
\alias{CellLabeler-class}
\title{Each CellLabeler object has a number of slots which store information. Key slots to access are listed below.}
\description{
Each CellLabeler object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{counts}}{The raw expression count matrix}

\item{\code{data}}{Normalized gene matrix}

\item{\code{scale.data}}{Scaled gene matrix}

\item{\code{meta.data}}{A cell-level metadata dataframe}

\item{\code{ude}}{A list of CellLabeler detected uniquely differential expression genes}

\item{\code{ModelFits}}{A list of model fitting results in ude detection}

\item{\code{ModelScores}}{A list of aggragated gene scores for each potential cell type}

\item{\code{ModelMarkers}}{A dataframe of CellLabeler identified marker genes overlapped between clusters and potential cell type}

\item{\code{prediction}}{A CellLabeler prediction dataframe}

\item{\code{num.core}}{The number of core used in the package}
}}

