% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLabeler.R
\name{celllabeler.CellLabeler}
\alias{celllabeler.CellLabeler}
\title{CellLabeler:}
\usage{
\method{celllabeler}{CellLabeler}(
  object,
  features = NULL,
  markers = NULL,
  sample.var = "sample",
  cluster.var = "cluster",
  similar.gene = 20,
  similar.pct = 0.8,
  lfc = 0.5,
  max.genes = 100,
  min.ccells = 20,
  up.thr = 0.9,
  mod = "combine",
  num.core = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A CellLabeler object}

\item{features}{Feature names. Default is NULL, then all features will be used.}

\item{sample.var}{Column names of sample information in metadata.}

\item{cluster.var}{Column names of cluster information in metadata.}

\item{similar.gene}{Integer value as the number of top expressed genes in considering combining clusters before DE test}

\item{similar.pct}{Numeric value in \link{0,1} as the threshold to combine clusters before DE test when either two cluster have enough overlapped highly expressed genes. Default is 0.8.}

\item{lfc}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between the interested clusters and the rest others. Default is 0.5.}

\item{max.genes}{Maximum number of highly expressed genes in each cluster required. Default is 100.}

\item{min.ccells}{Minimum number of cells in each cluster required. Default is 10.}

\item{up.thr}{Numberic value in \link{0,1} as the proportion of the remaining clusters in which a gene is upregulated in the interested group compared with it. Highter, the more strict. Default 0.9.}

\item{mod}{Character as "combine" or "all", specifying if combining similar clusters or not before DE test. Default is "combine"}

\item{num.core}{Number of cores in multi-threaded running. Default is 1.}
}
\value{
object A CellLabeler object
}
\description{
CellLabeler:
}
\author{
Jin Ning, Shiquan Sun
}
