% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLabeler.R
\name{celllabeler}
\alias{celllabeler}
\title{Unique marker gene detection and automatic cell type annotation}
\usage{
celllabeler(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Unique marker gene detection and automatic cell type annotation
}
\examples{

data(exampledata)
meta.data = data.frame(sample = sample.id, cluster = cluster.id, row.names = colnames(counts))
obj = CreateCellLabelerObject(counts = counts, meta.data = meta.data)
obj = celllabeler(obj, sample.var = "sample", cluster.var = "cluster", markers = markers)

}
\concept{data-access}
