% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{CombinePValues}
\alias{CombinePValues}
\title{Cauchy combination rules}
\usage{
CombinePValues(pvalues, weights = NULL)
}
\arguments{
\item{pvalues}{A n x m pvalue matrix; n genes, m permutations}

\item{weights}{A n x m weighs matrix for each genes in each permutation}
}
\value{
Returns a n-length combined pvalues
}
\description{
Cauchy combination rules
}
