% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ude.R
\name{ComputePCT}
\alias{ComputePCT}
\title{Compute gene expression percent of single-cell RNA data.}
\usage{
ComputePCT(cnt_data, cluster.id = NULL, sub.genes = NULL)
}
\arguments{
\item{cnt_data}{Gene expression raw count matrix.}

\item{cluster.id}{Character vector of cluster information in metadata.}

\item{sub.genes}{Subset of gene names.Default is NULL, then all genes will be used.}
}
\description{
Compute gene expression percent of single-cell RNA data.
}
